categories working 100%

+++ app/Http/Controllers/AdminUsersController.php
@@ -148,7 +148,11 @@ class AdminUsersController extends Controller
     {
         //
         $user = User::findOrFail($id);
-        unlink(public_path() . "../../../"  . $user->photo->file);
+
+        if($user->photo){
+            unlink(public_path() . "../../../"  . $user->photo->file); 
+        }
+
         $user->delete();
         
         // User::findOrFail($id)->delete();
